fix errors, create more pure functions from handle_connection

+++ src/http_def.c
@@ -25,7 +25,7 @@ const static struct
 
 REQUEST_TYPE string_to_request_type(const char *str)
 {
-    for (int i = 0; i < (sizeof(__REQUEST_TYPE_STRUCT) / sizeof(__REQUEST_TYPE_STRUCT[0])); i++)
+    for (unsigned long i = 0; i < (sizeof(__REQUEST_TYPE_STRUCT) / sizeof(__REQUEST_TYPE_STRUCT[0])); i++)
     {
         if (!strcmp(str, __REQUEST_TYPE_STRUCT[i].str))
         {
@@ -37,7 +37,7 @@ REQUEST_TYPE string_to_request_type(const char *str)
 
 const char *request_type_to_string(REQUEST_TYPE type)
 {
-    for (int i = 0; i < (sizeof(__REQUEST_TYPE_STRUCT) / sizeof(__REQUEST_TYPE_STRUCT[0])); i++)
+    for (unsigned long i = 0; i < (sizeof(__REQUEST_TYPE_STRUCT) / sizeof(__REQUEST_TYPE_STRUCT[0])); i++)
     {
         if (type == __REQUEST_TYPE_STRUCT[i].request_type)
         {
@@ -49,7 +49,7 @@ const char *request_type_to_string(REQUEST_TYPE type)
 
 const char *http_code_to_message(int code)
 {
-    for (int i = 0; i < (sizeof(__HTTP_CODE_MESSAGES) / sizeof(__HTTP_CODE_MESSAGES[0])); i++)
+    for (unsigned long i = 0; i < (sizeof(__HTTP_CODE_MESSAGES) / sizeof(__HTTP_CODE_MESSAGES[0])); i++)
     {
         if (code == __HTTP_CODE_MESSAGES[i].code)
         {
